Split test_dataset.py completely into test_dataset_{consumption, map, all_to_all}.py
