+++ python/ray/data/tests/test_transform_pyarrow.py
@@ -1,7 +1,12 @@
+import os
+
 import numpy as np
+import pandas as pd
 import pyarrow as pa
 import pytest
 
+import ray
+from ray.data.block import BlockAccessor
 from ray.data.extensions import (
     ArrowTensorArray,
     ArrowTensorType,
@@ -284,6 +289,109 @@ def test_unify_schemas():
             ("col_var_tensor", ArrowVariableShapedTensorType(pa.int16(), 5)),
         ]
     )
+
+
+def test_arrow_block_select():
+    df = pd.DataFrame({"one": [10, 11, 12], "two": [11, 12, 13], "three": [14, 15, 16]})
+    table = pa.Table.from_pandas(df)
+    block_accessor = BlockAccessor.for_block(table)
+
+    block = block_accessor.select(["two"])
+    assert block.schema == pa.schema([("two", pa.int64())])
+    assert block.to_pandas().equals(df[["two"]])
+
+    block = block_accessor.select(["two", "one"])
+    assert block.schema == pa.schema([("two", pa.int64()), ("one", pa.int64())])
+    assert block.to_pandas().equals(df[["two", "one"]])
+
+    with pytest.raises(ValueError):
+        block = block_accessor.select([lambda x: x % 3, "two"])
+
+
+def test_arrow_block_slice_copy():
+    # Test that ArrowBlock slicing properly copies the underlying Arrow
+    # table.
+    def check_for_copy(table1, table2, a, b, is_copy):
+        expected_slice = table1.slice(a, b - a)
+        assert table2.equals(expected_slice)
+        assert table2.schema == table1.schema
+        assert table1.num_columns == table2.num_columns
+        for col1, col2 in zip(table1.columns, table2.columns):
+            assert col1.num_chunks == col2.num_chunks
+            for chunk1, chunk2 in zip(col1.chunks, col2.chunks):
+                bufs1 = chunk1.buffers()
+                bufs2 = chunk2.buffers()
+                expected_offset = 0 if is_copy else a
+                assert chunk2.offset == expected_offset
+                assert len(chunk2) == b - a
+                if is_copy:
+                    assert bufs2[1].address != bufs1[1].address
+                else:
+                    assert bufs2[1].address == bufs1[1].address
+
+    n = 20
+    df = pd.DataFrame(
+        {"one": list(range(n)), "two": ["a"] * n, "three": [np.nan] + [1.5] * (n - 1)}
+    )
+    table = pa.Table.from_pandas(df)
+    a, b = 5, 10
+    block_accessor = BlockAccessor.for_block(table)
+
+    # Test with copy.
+    table2 = block_accessor.slice(a, b, True)
+    check_for_copy(table, table2, a, b, is_copy=True)
+
+    # Test without copy.
+    table2 = block_accessor.slice(a, b, False)
+    check_for_copy(table, table2, a, b, is_copy=False)
+
+
+def test_arrow_block_slice_copy_empty():
+    # Test that ArrowBlock slicing properly copies the underlying Arrow
+    # table when the table is empty.
+    df = pd.DataFrame({"one": []})
+    table = pa.Table.from_pandas(df)
+    a, b = 0, 0
+    expected_slice = table.slice(a, b - a)
+    block_accessor = BlockAccessor.for_block(table)
+
+    # Test with copy.
+    table2 = block_accessor.slice(a, b, True)
+    assert table2.equals(expected_slice)
+    assert table2.schema == table.schema
+    assert table2.num_rows == 0
+
+    # Test without copy.
+    table2 = block_accessor.slice(a, b, False)
+    assert table2.equals(expected_slice)
+    assert table2.schema == table.schema
+    assert table2.num_rows == 0
+
+
+def test_convert_to_pyarrow(ray_start_regular_shared, tmp_path):
+    ds = ray.data.range(100)
+    assert ds.to_dask().sum().compute()[0] == 4950
+    path = os.path.join(tmp_path, "test_parquet_dir")
+    os.mkdir(path)
+    ds.write_parquet(path)
+    assert ray.data.read_parquet(path).count() == 100
+
+
+def test_pyarrow(ray_start_regular_shared):
+    ds = ray.data.range_table(5)
+    assert ds.map(lambda x: {"b": x["value"] + 2}).take() == [
+        {"b": 2},
+        {"b": 3},
+        {"b": 4},
+        {"b": 5},
+        {"b": 6},
+    ]
+    assert ds.map(lambda x: {"b": x["value"] + 2}).filter(
+        lambda x: x["b"] % 2 == 0
+    ).take() == [{"b": 2}, {"b": 4}, {"b": 6}]
+    assert ds.filter(lambda x: x["value"] == 0).flat_map(
+        lambda x: [{"b": x["value"] + 2}, {"b": x["value"] + 20}]
+    ).take() == [{"b": 2}, {"b": 20}]
 
 
 if __name__ == "__main__":
